k1c: Force 64 bits otherwise it will be signed extended
Summary:
If we don't force 64 bits the value 0x80000000 will be signed
extended as 0xffffffff80000000.

Test Plan: None

Reviewers: O51 linux, cleger

Differential Revision: https://phab.kalray.eu/D79

+++ arch/k1c/include/asm/sfr.h
@@ -54,9 +54,9 @@ static inline void
 k1c_sfr_clear_bit(unsigned char sfr, unsigned char bit)
 {
 	if (bit < 32)
-		wfxl(sfr, 1 << bit);
+		wfxl(sfr, (uint64_t) 1 << bit);
 	else
-		wfxm(sfr, 1 << (bit - 32));
+		wfxm(sfr, (uint64_t) 1 << (bit - 32));
 }
 
 #define k1c_sfr_set(_sfr, _val)	__builtin_k1_set(_sfr, _val)
